OFNumberTests: Fix type mismatch
FossilOrigin-Name: 14bc00f1180017fdd7b78343783030056beb6f4708d05129e299a008d87f3462

+++ new_tests/OFNumberTests.m
@@ -21,7 +21,7 @@
 @interface OFNumberTests: OTTestCase
 @end
 
-extern unsigned long OFHashSeed;
+extern unsigned long long OFHashSeed;
 
 @implementation OFNumberTests
 - (void)testIsEqual
